logistica_tracking/backend/requirements.txt
 Django==5.1.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
requests==2.32.3
PyYAML==6.0.3
setuptools==82.0.1
django-cors-headers==4.4.0